Almost working user/admin functionality

+++ public/js/admin.js
@@ -4,8 +4,8 @@ let users = [], companies = [], jobPosts = [];
 let usersUl, companiesUl, jobPostsDiv;
 
 async function initPage(e) {
-	users = getAllUsers();
-	companies = getAllCompanies();
+	users = await getAllApplicants();
+	companies = await getAllEmployers();
 	jobPosts = await getAllJobPosts();
 
 	usersUl = document.querySelector("#users");
@@ -26,28 +26,48 @@ async function initPage(e) {
 
 window.addEventListener("load", initPage);
 
-function usersClickListener(e) {
+async function usersClickListener(e) {
 	if(e.target.classList.contains("delete"))
-		deleteUser(e);
+		await deleteUser(e);
 }
 
-function deleteUser(e) {
+async function deleteUser(e) {
 	const userLi = e.target.parentElement.parentElement;
 	const idx = Array.prototype.indexOf.call(usersUl.children, userLi);
-	deleteAccount(users[idx]._id);
+	const deleted = await deleteAccount(users[idx]._id);
+	if(!deleted) {
+		alert('Something went wrong...');
+		return;
+	}
 	users.splice(idx, 1);
 	removeUserLi(userLi);
 }
 
-function companiesClickListener(e) {
+async function companiesClickListener(e) {
 	if(e.target.classList.contains("delete"))
-		deleteCompany(e);
+		await deleteCompany(e);
 }
 
-function deleteCompany(e) {
+async function deleteCompany(e) {
 	const companyLi = e.target.parentElement.parentElement;
 	const idx = Array.prototype.indexOf.call(companiesUl.children, companyLi);
-	deleteAccount(companies[idx]._id);
+	const postsDeleted = await deleteCompanyPosts(companies[idx]._id);
+	if(!postsDeleted) {
+		alert('Something went wrong...');
+		return;
+	}
+	jobPosts.filter(p => {
+		return p.creator.toString() === companies[idx]._id.toString();
+	}).forEach(p => {
+		const i = Array.prototype.indexOf.call(jobPosts, p)
+		jobPosts.splice(i, 1);
+		removePostDiv(jobPostsDiv.children[i]);
+	});
+	const deleted = await deleteAccount(companies[idx]._id);
+	if(!deleted) {
+		alert('Something went wrong...');
+		return;
+	}
 	companies.splice(idx, 1);
 	removeCompanyLi(companyLi);
 }
@@ -61,7 +81,10 @@ async function deletePost(e) {
 	const postDiv = e.target.parentElement.parentElement.parentElement;
 	const idx = Array.prototype.indexOf.call(jobPostsDiv.children, postDiv);
 	const deleted = await deleteJobPost(jobPosts[idx]._id);
-	if(!deleted) return;
+	if(!deleted) {
+		alert('Something went wrong...');
+		return;
+	}
 	jobPosts.splice(idx, 1);
 	removePostDiv(postDiv);
 }
@@ -70,9 +93,9 @@ async function deletePost(e) {
 function renderAccount(account, accountUl) {
 	const accountLi = document.createElement("li");
 	accountLi.className = "list-group-item d-flex justify-content-between align-items-center";
-	accountLi.innerHTML = `<span>${account}</span>
+	accountLi.innerHTML = `<span>${account.username}</span>
       <span>
-        <a href="${getProfileLink(account)}"><button type="button" class="btn btn-success rounded-circle" data-toggle="tooltip" data-placement="bottom" title="View profile"><i class="far fa-eye"></i></button></a>
+        <a href="/u/${account._id}"><button type="button" class="btn btn-success rounded-circle" data-toggle="tooltip" data-placement="bottom" title="View profile"><i class="far fa-eye"></i></button></a>
         <!--<button type="button" class="btn btn-primary rounded-circle" data-toggle="tooltip" data-placement="bottom" title="Edit details of this user"><i class="fas fa-pen"></i></button>-->
         <button type="button" class="delete btn btn-danger rounded-circle" data-toggle="tooltip" data-placement="bottom" title="Delete account"><i class="no-click fas fa-times"></i></button>
       </span>`;
@@ -107,28 +130,92 @@ function removePostDiv(postDiv) {
 }
 
 /** BACKEND MANIPULATING FUNCTIONS */
-function deleteAccount(id) {
-	// Delete account from db
+async function deleteAccount(id) {
+	// Should delete the given job post from db
 	// code below requires server call
-	return true;
+	const url = `/users/${id}`;
+	const request = new Request(url, {
+	        method: 'DELETE',
+	        body: JSON.stringify({}),
+	        headers: {
+	            'Accept': 'application/json, text/plain, */*',
+	            'Content-Type': 'application/json'
+	        },
+	    });
+
+	const res = await fetch(request);
+	return res.status === 200;
 }
 
-function getAllUsers() {
+async function deleteCompanyPosts(id) {
+	const url = `/posts/${id}`;
+	const request = new Request(url, {
+	        method: 'DELETE',
+	        body: JSON.stringify({}),
+	        headers: {
+	            'Accept': 'application/json, text/plain, */*',
+	            'Content-Type': 'application/json'
+	        },
+	    });
+
+	const res = await fetch(request);
+	return res.status === 200;
+}
+
+async function getAllApplicants() {
 	// Get all users from db
 	// code below requires server call
 	// this function will be rewritten
 	// to only get a few users (pagination)
 	// from backend (when it is implemented)
-	return ["User 1", "User 2"];
+	//return ["User 1", "User 2"];
+	const url = `/applicants`;
+    // The data we are going to send in our request
+    // Create our request constructor with all the parameters we need
+    const request = new Request(url, {
+        method: 'get',
+        headers: {
+            'Accept': 'application/json, text/plain, */*',
+            'Content-Type': 'application/json'
+        },
+    });
+    const res = await fetch(request)
+    const json = await res.json();
+    if(json.error) {
+		alert(json.error);
+	} else if(json.redirect) {
+		window.location.href = json.redirect;
+	}
+
+    return json;
 }
 
-function getAllCompanies() {
+async function getAllEmployers() {
 	// Get all companies from db
 	// code below requires server call
 	// this function will be rewritten
 	// to only get a few companies (pagination)
 	// from backend (when it is implemented)
-	return ["Company 1", "Company 2"];
+	//return ["Company 1", "Company 2"];
+	const url = `/employers`;
+    // The data we are going to send in our request
+    // Create our request constructor with all the parameters we need
+    const request = new Request(url, {
+        method: 'get',
+        headers: {
+            'Accept': 'application/json, text/plain, */*',
+            'Content-Type': 'application/json'
+        },
+    });
+    const res = await fetch(request)
+    const json = await res.json();
+    if(json.error) {
+		alert(json.error);
+	} else if(json.redirect) {
+		window.location.href = json.redirect;
+	}
+
+    return json;
 }
 
 function getProfileLink(user) {
